packages/electoral-system-api/requirements.txt
alembic==1.13.1
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
typing_extensions==4.12.0
Werkzeug==3.0.3
